+++ pages/cdc_abmcostos.php
@@ -278,7 +278,18 @@ desired effect
                             </div>
                             <div class="row">
                                 
-                                <div class="col-md-2"></div>          
+                                <div class="col-md-2">
+                                    <div class="form-group">
+                                        <label for="oportunidad">#CERES</label>
+                                        <input type="text" class="form-control" name="oportunidad"  id='modal-abm-costos-oportunidad' required value="<?= ($id_planilla ? $planilla_costeo['oportunidad_comercial'] : "") ?>">
+                                    </div>                                       
+                                </div>          
+                                <div class="col-md-2">
+                                    <div class="form-group">
+                                        <label for="ss">SS</label>
+                                        <input placeholder="SS...." type="text" class="form-control" name="ss"  id='modal-abm-costos-ss' required value="<?= ($id_planilla ? $planilla_costeo['solicitud_servicio'] : "") ?>">
+                                    </div>                                     
+                                </div>          
                                 <div class="col-md-2">
                                     <div class="form-group">
                                         <label for="meses">Meses de cto.</label>